Add hidden `pcb embed-step` command
