+++ crates/pcb-kicad/Cargo.toml
@@ -9,14 +9,20 @@ description = "KiCad integration for PCB design tools"
 
 [dependencies]
 anyhow = { workspace = true }
+atomicwrites = { workspace = true }
+base64 = { workspace = true }
 dirs = { workspace = true }
+pcb-sexpr = { workspace = true }
 tempfile = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 typify = { workspace = true }
 chrono = { workspace = true }
 regress = { workspace = true }
+regex = { workspace = true }
 log = { workspace = true }
 pcb-command-runner = { workspace = true }
 pcb-zen-core = { workspace = true }
+sha2 = { workspace = true }
 starlark = { workspace = true }
+zstd = { workspace = true }
